tests for message model and views

+++ unit26/warbler/app.py
@@ -379,6 +379,11 @@ def messages_destroy(message_id):
 
     msg = Message.query.get(message_id)
 
+    if msg.user_id != g.user.id:
+        flash("Access unauthorized.", "danger")
+
+        return redirect('/')
+
     db.session.delete(msg)
     db.session.commit()
